+++ src/validations/prefab/componentOption.ts
@@ -20,6 +20,13 @@ const refSchema = Joi.when('type', {
   otherwise: Joi.object({ id: Joi.string() }),
 });
 
+const optionConditionSchema = Joi.object({
+  type: Joi.string().valid(...CONDITION_TYPE),
+  option: Joi.string(),
+  comparator: Joi.string().valid(...COMPARATORS),
+  value: Joi.any(),
+});
+
 const optionConfigurationSchemaBase = {
   apiVersion: Joi.string()
     .pattern(/^v[\d]{1,}/)
@@ -33,12 +40,7 @@ const optionConfigurationSchemaBase = {
   allowedTypes: Joi.array().items(Joi.string()),
   as: Joi.string().valid(...CONFIGURATION_AS),
   component: Joi.string(),
-  condition: Joi.object({
-    type: Joi.string().valid(...CONDITION_TYPE),
-    option: Joi.string(),
-    comparator: Joi.string().valid(...COMPARATORS),
-    value: Joi.any(),
-  }),
+  condition: optionConditionSchema,
   disabled: Joi.boolean(),
   dataType: Joi.string(),
   dependsOn: Joi.string(),
@@ -83,20 +85,6 @@ const optionConfigurationSchema = Joi.when('type', {
 
   .default({});
 
-const linkedOptionValueSchema = Joi.object({
-  ref: Joi.object({
-    componentId: Joi.string().required(),
-    optionId: Joi.string().required(),
-  }),
-});
-
-export const linkedOptionSchema = Joi.object({
-  key: Joi.string().required(),
-  label: Joi.string().allow(''),
-  type: Joi.string().valid('LINKED_OPTION').required(),
-  value: linkedOptionValueSchema,
-});
-
 export const optionSchema = Joi.object({
   label: Joi.string().required(),
   key: Joi.string().required(),
@@ -110,4 +98,11 @@ export const optionSchema = Joi.object({
     otherwise: Joi.any(),
   }),
   ref: refSchema,
+});
+
+export const optionCategorySchema = Joi.object({
+  label: Joi.string().required(),
+  expanded: Joi.boolean(),
+  members: Joi.array().items(Joi.string()).min(1).required(),
+  condition: optionConditionSchema,
 });